fix(bookings): move overlap-meta helpers out of the availability route file
Second instance of the route-export rule: Next route modules may only
export HTTP-method fields. Helpers and their types now live in
lib/booking/overlap-meta.

Part of #1018.
